Clarify borrowed file server-side copy restrictions
